Jenkinsfile: take out rpmdeplint step for now
This is intermittently failing due to problems with the Fedora download
server. And we also shouldn't be pulling directly from
dl.fedoraproject.org anyway, we should be using a mirror.

The proper solution would be to have rpmdeplint use the mirrorlist or
metalink. That's not supported currently, rpmdeplint RFE is:

    https://bugzilla.redhat.com/show_bug.cgi?id=1454525