exclude conf/settings.py from published tarballs
The waiverdb-0.2.1 tarball I published accidentally included a 0-byte
conf/settings.py because that file was (uncommitted) in my working tree
when I built the tarball. Moreover I didn't notice because we explicitly
gitignore'd that file because we *expect* everyone to have to it in
their working tree.

So let's ensure that the setuptools manifest won't pick it up in future.
A nicer solution might be to switch to setuptools-git to ensure the
tarball *never* includes random uncommitted junk in it...