don't fail the RPM build on Sphinx warnings
Now that we are using sphinxcontrib-issuetracker, some warnings are
unfortunately expected when we build inside Koji, because it won't be
able to reach the network:

    WARNING: GET https://pagure.io/api/0/waiverdb/issue/73 failed with error: ...

We can just ignore them. The Jenkins pipeline will still build the docs
with SPHINXOPTS=-W so we will still notice any *other* warnings that
crop up.