database migrations should be a pre-deployment hook
... not a post-start pod hook for the web app. This way, we can see the
logs of the migration pod, and if the migration fails the deployment
will be aborted.

The database migration pod has to first wait for the Postgres pod to
finish starting up, because in a freshly created environment (as in the
tests) the migration pod is started concurrently with the Postgres pod.

Fixes #121.