run mock build with networking enabled
By default, networking is not enabled in a mock chroot. This causes sphinx to
timeout while trying to query pagure.io about issues referenced in the docs.
This significantly slows down the mock build. By enabling networking, we
avoid the timeouts and speed up the build. Reproduceability is not a concern,
since this is a throwaway test build anyway.