make waiver.subject column type JSONB
The waiver.subject column was previously using the Postgres JSON column
type, but this does not support indexing or equality comparison. Use the
newer JSONB column type instead, which supports btree indexes and can be
efficiently queried with equality comparisons.

Also, the database migration which created the waiver.subject
accidentally created the column as TEXT not JSON. This is handled in
a new migration with `USING subject::jsonb` which causes Postgres to
parse the column values into JSON when converting the column.

Dropped the dependency on sqlalchemy-utils since we are no longer using
it for anything.

Fixes #129.